fix: reduce concurrent Auth0 token refresh races
