+++ packages/utils/package.json
@@ -23,6 +23,7 @@
     "@date-fns/tz": "^1.2.0",
     "date-fns": "^4.1.0",
     "html-to-text": "^9.0.5",
+    "jose": "^6.0.11",
     "slugify": "^1.6.6",
     "zod": "^3.25.76"
   }
